backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
python-socketio==5.11.0
APScheduler==3.10.4
requests==2.31.0
gunicorn==21.2.0
eventlet==0.35.2
ffmpeg-python==0.2.0

